Refactor: 도서 수정/삭제 코드 분리
